refactor: break down migrate tasks
